"check_style_score -> 2 of 4 passed. style score: 0/1 "

+++ ADS1-assignments/m3/3-sum/Solution.java
@@ -3,7 +3,7 @@ class ThreeSum {
 	ThreeSum() {
 
 	}
-	public int calculate(int[] array) {
+	public int calculate(long[] array) {
 		int count = 0;
 		for (int i = 0; i < array.length; i++) {
 			for (int j = i + 1; j < array.length; j++) {
@@ -21,7 +21,7 @@ final class Solution {
 	public static void main(final String[] args) {
 		Scanner scan = new Scanner(System.in);
 		int length = scan.nextInt();
-		int[] array = new int[length];
+		long[] array = new long[length];
 		ThreeSum obj = new ThreeSum();
 		for (int i = 0; i < length; i++) {
 			array[i] = scan.nextInt();
